final-project/110550014_final/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch
torchvision
torchaudio
transformers
datasets
pillow
tqdm